Support 'BAZELISK_NOJDK' in Bazelisk Go
