+++ configs/BUILD
@@ -1,0 +1,10 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
+
+go_library(
+    name = "go_default_library",
+    srcs = [
+        "config.go",
+    ],
+    importpath = "github.com/bazelbuild/bazelisk/configs",
+    visibility = ["//visibility:public"],
+)
