+++ platforms/BUILD
@@ -6,6 +6,7 @@ go_library(
     importpath = "github.com/bazelbuild/bazelisk/platforms",
     visibility = ["//visibility:public"],
     deps = [
+        "//configs:go_default_library",
         "//versions:go_default_library",
         "@com_github_hashicorp_go_version//:go_default_library",
     ],
